+++ src/routes/Router.tsx
@@ -2,7 +2,7 @@ import { BrowserRouter, Route, Routes } from 'react-router-dom'
 import MainLayout from '@/component/layout/MainLayout'
 import Home from '@/page/Home'
 import AdminHome from '@/page/AdminHome'
-import NotFound from '@/page/NotFount'
+import NotFound from '@/page/NotFound.tsx'
 import User from '@/page/user/User'
 import Login from '@/page/user/Login'
 import AdminLogin from '@/page/user/AdminLogin'
